+++ lib/services/storage_persist_service_stub.dart
@@ -1,5 +1,3 @@
-import 'package:flutter/material.dart';
-
 class StoragePersistService {
-  static Future<void> requestIfNeeded(BuildContext context) async {}
+  static Future<void> requestIfNeeded() async {}
 }